+++ public/app/core/components/AppChrome/ReportIssueButton/types.ts
@@ -1,0 +1,11 @@
+export type FeedbackFormData = {
+  message: string;
+  screenshot: string;
+  imageType: string;
+  reporterEmail: string;
+  accessChecked: boolean;
+  contactChecked: boolean;
+  width: number;
+  height: number;
+  bitmap: HTMLImageElement;
+};
