+++ public/app/core/components/AppChrome/ReportIssueButton/utils.ts
@@ -1,0 +1,3 @@
+export function isCanvas(obj: HTMLCanvasElement | HTMLElement): obj is HTMLCanvasElement {
+  return obj && obj.tagName === 'CANVAS';
+}
